File upload and download

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator'); //importing validationResult function from express-validator
 
 const Product = require('../models/product');
@@ -20,14 +21,31 @@ exports.getAddProduct = (req,res,next)=>{
 
 exports.postAddProduct = (req,res,next)=>{ 
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
+    const genre = req.body.genre;
     const description = req.body.description;
-    // console.log(req);
+    // console.log(imageUrl);
     // console.log(errors.array());
+
+    if(!image){
+      return res.status(422).render('admin/edit-product', { 
+        pageTitle: 'Add Product',
+        path: '/admin/add-product',
+        editing : false,
+        hasError: true,
+        product: {
+          title: title,
+          price: price,
+          genre: genre, 
+          description: description
+        },
+        errorMessage: 'Attached file is not an image.',
+        validationErrors: []
+      }); 
+    }
     
     const errors = validationResult(req);
-
     if(!errors.isEmpty()){
       console.log(errors.array());
       return res.status(422).render('admin/edit-product', { 
@@ -38,6 +56,7 @@ exports.postAddProduct = (req,res,next)=>{
         product: {
           title: title,
           price:price, 
+          genre: genre,
           description:description
         },
         errorMessage: errors.array()[0].msg,
@@ -45,10 +64,13 @@ exports.postAddProduct = (req,res,next)=>{
       });
     }
 
+    const imageUrl = image.path;
+
     const product = new Product({
       title: title,
       price:price, 
       description:description, 
+      genre: genre,
       imageUrl:imageUrl,
       userId: req.user
     }) //mapping
@@ -97,7 +119,8 @@ exports.postEditProduct = (req,res,next)=>{
       const prodId = req.body.productId;
       const updatedTitle = req.body.title;
       const updatedPrice = req.body.price;
-      const updatedImageUrl = req.body.imageUrl;
+      const updatedGenre = req.body.genre;
+      const image = req.file;
       const updatedDesc = req.body.description;
 
       const errors = validationResult(req);
@@ -112,7 +135,7 @@ exports.postEditProduct = (req,res,next)=>{
           title: updatedTitle,
           price: updatedPrice, 
           description: updatedDesc, 
-          imageUrl: updatedImageUrl,
+          genre: updatedGenre,
           _id: prodId
         },
         errorMessage: errors.array()[0].msg,
@@ -129,7 +152,11 @@ exports.postEditProduct = (req,res,next)=>{
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        product.genre = updatedGenre;
+        if(image){
+          fileHelper.deleteFile(product.imageUrl);
+          product.imageUrl = image.path;
+        }
         return product.save() .then(result => {
           console.log('UPDATED PRODUCT!!');
           res.redirect('/admin/products');
@@ -164,7 +191,16 @@ exports.getProducts = (req,res,next)=>{
 
 exports.deleteProduct = (req,res,next)=>{
       const prodId = req.params.productId;
-      Product.deleteOne({_id: prodId, userId: req.user._id}) //this removes the document with given prodId
+
+      Product.findById(prodId)
+      .then(product => {
+        if(!product)
+        {
+          return next(new Error('Product not found.'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id}) //this removes the document with given prodId
+      })
       .then(() => {
       console.log('DESTROYED PRODUCT!!');
       res.status(200).json({message: 'Success!'});
